Moved the build_group_transports to tests/QE
